Rename currents submodule to data
