+++ src/ship_routing/cost.py
@@ -1,4 +1,4 @@
-from .currents import select_data_for_leg
+from .data import select_data_for_leg
 
 import numpy as np
