+++ tests/test_algorihms.py
@@ -8,7 +8,7 @@ from ship_routing.algorithms import (
 )
 
 from ship_routing.core import Route, WayPoint
-from ship_routing.currents import load_currents
+from ship_routing.data import load_currents
 
 import numpy as np
 import pytest
